+++ src/ports/vscode/extra/CHANGELOG.md
@@ -56,7 +56,7 @@
 ![image](https://github.com/daltonmenezes/assets/blob/master/images/aura-theme/changelogs/v1.3.0/function-names.png?raw=true)
 
 ## 1.2.0 (August 21, 2020)
-- Add title bar colors - [#8](https://github.com/daltonmenezes/aura-theme/issues/8)
+- Add title bar colors - [#8](https://github.com/yukiumi13/aura-theme-2026/issues/8)
 
 ## 1.1.1 (August 19, 2020)
 - Update text selection background on inputs to be more clearer